Fix heartbeat spamming
In Windows, Path.Combine adds a back slash instead of a forward slash, making the heartbeat entity different from the entity in _lastHeartbeat response. Then it's not possible to skip any heartbeat. I tested and the server changes all backslashes into a forward slash. (eg: "\" to "/" and "\\\\\\\" to "/"). Application.dataPath and currentScene look ok because they use "/" in the editor.
Also added [Serializable] to the structs because _lastHeartbeat was not being set, I believe it is required for JSON.